confirm delete

+++ frontend/src/screens/UserListScreen.js
@@ -29,7 +29,9 @@ const UserListScreen = ({ history }) => {
     }, [dispatch, history, successDelete, userInfo]);
 
     const deleteHandler = (id) => {
-        dispatch(deleteUser(id))
+        if(window.confirm('Are you sure you want to delete this user?')){
+            dispatch(deleteUser(id))
+        }
     };
